重构CustomDictionary，支持多实例 https://github.com/hankcs/HanLP/issues/1339
