+++ src/main/java/com/hankcs/hanlp/seg/Viterbi/ViterbiSegment.java
@@ -40,11 +40,8 @@ import static com.hankcs.hanlp.utility.Predefine.logger;
  */
 public class ViterbiSegment extends WordBasedSegment
 {
-    private DoubleArrayTrie<CoreDictionary.Attribute> dat;
-
     public ViterbiSegment()
     {
-        this.dat = CustomDictionary.dat;
     }
 
     /**
@@ -66,12 +63,12 @@ public class ViterbiSegment extends WordBasedSegment
 
     public DoubleArrayTrie<CoreDictionary.Attribute> getDat()
     {
-        return dat;
+        return customDictionary.dat;
     }
 
     public void setDat(DoubleArrayTrie<CoreDictionary.Attribute> dat)
     {
-        this.dat = dat;
+        this.customDictionary.dat = dat;
     }
 
     @Override
@@ -94,8 +91,8 @@ public class ViterbiSegment extends WordBasedSegment
         if (config.useCustomDictionary)
         {
             if (config.indexMode > 0)
-                combineByCustomDictionary(vertexList, this.dat, wordNetAll);
-            else combineByCustomDictionary(vertexList, this.dat);
+                combineByCustomDictionary(vertexList, customDictionary.dat, wordNetAll);
+            else combineByCustomDictionary(vertexList, customDictionary.dat);
         }
 
         if (HanLP.Config.DEBUG)
@@ -214,10 +211,7 @@ public class ViterbiSegment extends WordBasedSegment
         File file = new File(mainPath);
         mainPath = file.getParent() + "/" + Math.abs(combinePath.toString().hashCode());
         mainPath = mainPath.replace("\\", "/");
-        if (CustomDictionary.loadMainDictionary(mainPath, path, dat, isCache))
-        {
-            this.setDat(dat);
-        }
+        customDictionary.loadMainDictionary(mainPath, path, dat, isCache);
     }
 
     /**
