+++ src/main/java/com/hankcs/hanlp/utility/LexiconUtility.java
@@ -83,12 +83,12 @@ public class LexiconUtility
         if (attribute == null) return false;
 
         if (CoreDictionary.trie.set(word, attribute)) return true;
-        if (CustomDictionary.dat.set(word, attribute)) return true;
-        if (CustomDictionary.trie == null)
+        if (CustomDictionary.DEFAULT.dat.set(word, attribute)) return true;
+        if (CustomDictionary.DEFAULT.trie == null)
         {
             CustomDictionary.add(word);
         }
-        CustomDictionary.trie.put(word, attribute);
+        CustomDictionary.DEFAULT.trie.put(word, attribute);
         return true;
     }
